Update OpenSSL + deps, default to http/1.1 when protocol cant be negotiated
