+++ esy.lock/opam/ppx_sexp_conv.v0.12.0/opam
@@ -14,7 +14,7 @@ depends: [
   "base"     {>= "v0.12" & < "v0.13"}
   "sexplib0" {>= "v0.12" & < "v0.13"}
   "dune"     {build & >= "1.5.1"}
-  "ppxlib"   {>= "0.5.0"}
+  "ppxlib"   {>= "0.5.0" & < "0.9.0"}
 ]
 synopsis: "[@@deriving] plugin to generate S-expression conversion functions"
 description: "
