--- esy.lock/opam/ppxfind.1.2/opam
@@ -1,26 +1,0 @@
-opam-version: "2.0"
-maintainer: "jeremie@dimino.org"
-authors: ["Jérémie Dimino"]
-license: "BSD3"
-homepage: "https://github.com/diml/ppxfind"
-bug-reports: "https://github.com/diml/ppxfind/issues"
-dev-repo: "git://github.com/diml/ppxfind.git"
-build: [
-  ["jbuilder" "build" "-p" name "-j" jobs]
-]
-depends: [
-  "ocaml" {>= "4.03.0"}
-  "jbuilder" {build & >= "1.0+beta16"}
-  "ocaml-migrate-parsetree"
-  "ocamlfind"
-]
-synopsis: "ocamlfind ppx tool"
-description: """
-Ppxfind is a small command line tool that among other things allows
-to use old style ppx rewriters with jbuilder."""
-url {
-  src:
-    "https://github.com/diml/ppxfind/releases/download/1.2/ppxfind-1.2.tbz"
-  checksum: "md5=3c6f81800bb816e190a64f9898481f82"
-}
-conflicts: [ "dune" {= "1.2.0" | = "1.2.1"} ]
