+++ esy.lock/opam/ppxfind.1.3/opam
@@ -1,0 +1,31 @@
+opam-version: "2.0"
+maintainer: "jeremie@dimino.org"
+authors: ["Jérémie Dimino"]
+license: "BSD3"
+homepage: "https://github.com/diml/ppxfind"
+bug-reports: "https://github.com/diml/ppxfind/issues"
+dev-repo: "git+https://github.com/diml/ppxfind.git"
+doc: "https://diml.github.io/ppxfind/"
+build: [
+  ["dune" "build" "-p" name "-j" jobs]
+]
+depends: [
+  "dune" {build & >= "1.0"}
+  "ocaml-migrate-parsetree"
+  "ocamlfind"
+  "ocaml" {>= "4.02.3"}
+]
+conflicts: [ "dune" {= "1.2.0" | = "1.2.1"} ]
+synopsis: "Tool combining ocamlfind and ppx"
+description: """
+Ppxfind is a small command line tool that among other things allows
+to use old style ppx rewriters with jbuilder.
+"""
+url {
+  src:
+    "https://github.com/diml/ppxfind/releases/download/1.3/ppxfind-1.3.tbz"
+  checksum: [
+    "sha256=d49db026d0e74212c4b475b4e628aa57508c7452a0682d8c96e80c130ab892e4"
+    "sha512=dd81bf5b3413f99a3c39f25e5e747f8d57d3bacac6a1fda92478af317e2d543294b4937982c94cf010978fcd71b1cfbfce1edbaa0c3d03973079296fda89689f"
+  ]
+}
