--- esy.lock/opam/result.1.3/opam
@@ -1,22 +1,0 @@
-opam-version: "2.0"
-maintainer: "opensource@janestreet.com"
-authors: ["Jane Street Group, LLC <opensource@janestreet.com>"]
-homepage: "https://github.com/janestreet/result"
-dev-repo: "git+https://github.com/janestreet/result.git"
-bug-reports: "https://github.com/janestreet/result/issues"
-license: "BSD3"
-build: [["jbuilder" "build" "-p" name "-j" jobs]]
-depends: [
-  "ocaml"
-  "jbuilder" {build & >= "1.0+beta11"}
-]
-synopsis: "Compatibility Result module"
-description: """
-Projects that want to use the new result type defined in OCaml >= 4.03
-while staying compatible with older version of OCaml should use the
-Result module defined in this library."""
-url {
-  src:
-    "https://github.com/janestreet/result/releases/download/1.3/result-1.3.tbz"
-  checksum: "md5=4beebefd41f7f899b6eeba7414e7ae01"
-}
