--- esy.lock/opam/syndic.1.6.0/opam
@@ -1,40 +1,0 @@
-opam-version: "2.0"
-name:         "syndic"
-maintainer:   "Romain Calascibetta <romain.calascibetta@gmail.com>"
-authors:      [ "Romain Calascibetta"
-                "Christophe Troestler" ]
-license:      "MIT"
-homepage:     "https://github.com/Cumulus/Syndic"
-dev-repo:     "git+https://github.com/Cumulus/Syndic.git"
-bug-reports:  "https://github.com/Cumulus/Syndic/issues"
-doc:          "https://cumulus.github.io/Syndic/"
-synopsis:     "RSS1, RSS2, Atom and OPML1 parsing"
-description: """
-Pure OCaml Library for parsing and writing various types of
-feeds and subscriber lists."""
-
-build: [
-  [ "dune" "subst" ] {pinned}
-  [ "dune" "build" "-p" name "-j" jobs ]
-  [ "dune" "runtest" "-p" name ] {with-test}
-]
-
-
-depends: [
-  "ocaml" {>= "4.04.0"}
-  "dune" {build}
-  "ptime"
-  "uri" {>= "1.9"}
-  "xmlm" {>= "1.2.0"}
-  "fmt" {with-test}
-  "ocurl" {with-test}
-  "fpath" {with-test}
-  "ocplib-json-typed" {with-test}
-  "base-unix" {with-test}
-  "jsonm" {with-test}
-]
-url {
-  src:
-    "https://github.com/Cumulus/Syndic/releases/download/v1.6.0/syndic-v1.6.0.tbz"
-  checksum: "md5=b587da170bfa5b5d2d931ce0105b22a4"
-}
