--- esy.lock/opam/yojson.1.5.0/opam
@@ -1,36 +1,0 @@
-opam-version: "2.0"
-maintainer: "martin@mjambon.com"
-authors: ["Martin Jambon"]
-homepage: "https://github.com/ocaml-community/yojson"
-bug-reports: "https://github.com/ocaml-community/yojson/issues"
-dev-repo: "git+https://github.com/ocaml-community/yojson.git"
-doc: "https://ocaml-community.github.io/yojson/"
-build: [
-  ["dune" "subst"] {pinned}
-  ["dune" "build" "-p" name "-j" jobs]
-]
-depends: [
-  "ocaml" {>= "4.02.3"}
-  "dune" {build}
-  "cppo" {build}
-  "easy-format"
-  "biniou" {>= "1.2.0"}
-]
-synopsis:
-  "Yojson is an optimized parsing and printing library for the JSON format"
-description: """
-Yojson is an optimized parsing and printing library for the JSON format.
-
-It addresses a few shortcomings of json-wheel including 2x speedup,
-polymorphic variants and optional syntax for tuples and variants.
-
-ydump is a pretty-printing command-line program provided with the
-yojson package.
-
-The program atdgen can be used to derive OCaml-JSON serializers and
-deserializers from type definitions."""
-url {
-  src:
-    "https://github.com/ocaml-community/yojson/releases/download/1.5.0/yojson-1.5.0.tbz"
-  checksum: "md5=d80de1bacdde292af42f7c78b323da7b"
-}
