+++ esy.lock/overrides/4d26755065a0ae5ee6bf67c12b4477cd/.travis.yml
@@ -1,0 +1,15 @@
+language: node_js
+node_js:
+- 8
+os:
+- linux
+- osx
+install:
+- npm install --global esy@0.3.0
+- esy install
+script:
+- travis_wait 40 esy build
+cache:
+    timeout: 360
+    directories:
+        - '$HOME/.esy/'
