+++ esy.lock/overrides/4d26755065a0ae5ee6bf67c12b4477cd/azure-pipelines.yml
@@ -1,0 +1,19 @@
+jobs:
+- template: .ci/build.yaml  # Template reference
+  parameters:
+    host: macOS
+    pool:
+      vmImage: 'macOS-10.13'
+
+- template: .ci/build.yaml  # Template reference
+  parameters:
+    host: Linux
+    pool:
+      vmImage: 'Ubuntu-16.04'
+
+- template: .ci/build.yaml  # Template reference
+  parameters:
+    host: Windows
+    pool:
+      vmImage: 'vs2017-win2016'
+    sign: true  # Extra step on Windows only
