+++ esy.lock/overrides/opam__s__conf_pkg_config_opam__c__1.1_opam_override/package.json
@@ -1,5 +1,11 @@
 {
+  "build": [
+    [
+      "pkg-config",
+      "--help"
+    ]
+  ],
   "dependencies": {
-    "yarn-pkg-config": "esy-ocaml/yarn-pkg-config#d488cd9321cd5036bd36ec96744ce78c5d45fc49"
+    "yarn-pkg-config": "esy-ocaml/yarn-pkg-config#cca65f99674ed2d954d28788edeb8c57fada5ed0"
   }
 }
