+++ esy.lock/overrides/opam__s__dune_opam__c__1.10.0_opam_override/package.json
@@ -1,0 +1,14 @@
+{
+  "build": [
+    [
+      "ocaml",
+      "bootstrap.ml"
+    ],
+    [
+      "./boot.exe",
+      "--release",
+      "-j",
+      "4"
+    ]
+  ]
+}
