--- esy.lock/overrides/opam__s__dune_opam__c__1.9.1_opam_override/package.json
@@ -1,14 +1,0 @@
-{
-  "build": [
-    [
-      "ocaml",
-      "bootstrap.ml"
-    ],
-    [
-      "./boot.exe",
-      "--release",
-      "-j",
-      "4"
-    ]
-  ]
-}
