--- esy.lock/overrides/opam__s__ppx__deriving_opam__c__4.2.1_opam_override/package.json
@@ -1,15 +1,0 @@
-{
-  "build": [
-    [
-      "bash",
-      "-c",
-      "#{os == 'windows' ? 'patch -p1 < ppx_deriving-4.2.1.patch' : 'true'}"
-    ],
-    [
-      "ocaml",
-      "pkg/build.ml",
-      "native=true",
-      "native-dynlink=true"
-    ]
-  ]
-}
